ftest/requirements.txt

cffi==1.15.1
channels==3.0.4
constantly==15.1.0
cryptography==39.0.2
daphne==3.0.2
Django==3.0.1
django-cors-headers==3.10.1
django-filter==21.1
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
paramiko==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==23.0.0
pytz==2022.7.1
semantic-version==2.10.0
service-identity==21.1.0
setuptools-rust==1.1.2
